move matrix and hierCluster code to plots/matrix subdir
