+++ client/mass/test/matrix.integration.spec.js
@@ -2,7 +2,7 @@ import tape from 'tape'
 import { termjson } from '../../test/testdata/termjson'
 import * as helpers from '../../test/front.helpers.js'
 import { sleep, detectLst, detectGte } from '../../test/test.helpers.js'
-import { getSortOptions } from '../../plots/matrix.sort.js'
+import { getSortOptions } from '../../plots/matrix/matrix.sort.js'
 import {
 	proteinChangingMutations,
 	truncatingMutations,
