+++ client/plots/matrix/matrix.cluster.js
