+++ client/plots/matrix/matrix.dom.js
