+++ client/plots/matrix/matrix.layout.js
@@ -1,4 +1,4 @@
-import { getCompInit, copyMerge, deepEqual } from '../rx'
+import { getCompInit, copyMerge, deepEqual } from '#rx'
 import { scaleLinear, scaleOrdinal } from 'd3-scale'
 import { schemeCategory10, interpolateReds, interpolateBlues } from 'd3-scale-chromatic'
 import { axisLeft, axisTop, axisRight, axisBottom } from 'd3-axis'
