+++ client/plots/matrix/matrix.serieses.js
@@ -1,5 +1,5 @@
 import { setCellProps, getEmptyCell, maySetEmptyCell, setGeneVariantCellProps } from './matrix.cells'
-import { TermTypes } from '../shared/terms.js'
+import { TermTypes } from '#shared/terms.js'
 
 export function getSerieses(data) {
 	const s = this.settings.matrix
