+++ client/tw/index.ts
@@ -1,3 +1,4 @@
 export * from './categorical'
 export * from './numeric'
 export * from './composite'
+export * from './TwRouter'
